[dart2wasm] Avoid duplicate checks in dynamic forwarders
In a forwarder for a method, check class IDs of all classes inheriting
the member and run the shape check once.

For example, in a this class hierarchy:

    class A {
      void printName() { print("A"); }
    }

    class B extends A {
      void printName() { print("B"); }
    }

    class C extends A {}

When generating method forwarder for `printName`, we currently generate
a shape check for all classes.

This this CL we now generate two shape checks: one for A and C (which
share the same `printName`), one for B.

Fixes #53880.