[dart2wasm] Remove array.new_data handling in global sections
As of 2025/05/14, `array.new_data` is not a constant instruction, so
can't be used in a global initializer.

Also, data count section needs to appear right before the code section.
Without `array.new_data` we don't need to generate it before the global
section.

Remove the code related to handling `array.new_data` in global sections.

When `array.new_data` becomes constant, we can generate the data count
section in its new place always, without having to specially handle it
when it's used in a global section.