[dart2wasm] Move async compiler utilities to state_machine library
This is the second part of the refactoring to make sync* and async
compilers to reuse code as much as possible.

Move exception handler and label target types from `async` to
`state_machine` library.

This is a change from
https://dart-review.googlesource.com/c/sdk/+/366663, moved to a spearate
CL to make revieweing easier.

The end goal is to share code generation for async and sync* code
generators.