[dart2wasm] Remove a duplicated loop in dart:convert
This removes the outer check in `List<int>` element check:

```
if (allowMalformed) {
   ... check code units in `List<int>`, write 0xFF for invalid bytes ...
} else {
   ... check code units in `List<int>`, throw for invalid bytes ...
}
```

and moves the `allowMalformed` check to the inside of the loop, when the
`int` value is negative or more then 255.

The idea is that `value < 0 || value > 255` will hold rarely, so it's OK
to have slow code inside the conditional.

On a side note, `allowMalformed` in UTF8 is use for two things:

1. To allow invalid bytes in `List<int>` argument.
2. To allow invalid UTF-8 encodings.

Both of these will happen rarely, but (1) will almost never happen as
any code that cares about performance will pass a `Uint8List`. So it
makes sense to assume that (1) will hold very rarely, and optimize based
on that.